it lives! finally deployed to streamlit could hooray! :)
