+++ streamlit_app.py
@@ -1,153 +1,13 @@
 """
 
 """
-import boto3
-import pandas as pd
-from io import StringIO
 import streamlit as st
-import plotly.express as px
-import plotly.graph_objects as go
+
+# This file can be kept minimal since Streamlit automatically handles page navigation
+st.title('Australian Powerlifting League Analysis')
+st.write('Please select a page from the left sidebar to view different analyses.')
 
 
-aws_access_key_id = st.secrets['AWS_ACCESS_KEY_ID']
-aws_secret_access_key = st.secrets['AWS_SECRET_ACCESS_KEY']
-
-
-s3 = boto3.client(
-    's3',
-    aws_access_key_id=aws_access_key_id,
-    aws_secret_access_key=aws_secret_access_key
-)
-
-bucket_name = 'open-powerlifting'
-object_key = 'open-powerlifting-australia.csv'
-
-obj = s3.get_object(Bucket=bucket_name, Key=object_key)
-
-dtype_dict = {
-    'State': 'string',
-    'ParentFederation': 'string',
-    'MeetState': 'string',
-    'Bench4kg': 'float'
-}
-
-
-def yes_no_to_bool(value):
-    """
-    converts "Yes" to True and empty values to False
-    :param value:
-    :return:
-    """
-    if value == 'Yes':
-        return True
-    else:
-        return False
-
-
-df = pd. read_csv(
-    StringIO(obj['Body'].read().decode('utf-8')),
-    dtype=dtype_dict,
-    converters={'Tested': yes_no_to_bool}
-)
-
-filtered_df = df[df['Federation'] == 'AusPL']
-filtered_df_copy = filtered_df.copy()
-filtered_df_copy['Date'] = pd.to_datetime(filtered_df_copy['Date'])
-filtered_df_copy['Year'] = filtered_df_copy['Date'].dt.year
-
-
-yearly_meet_counts = filtered_df_copy.groupby('Year')['MeetName'].nunique()
-# fig = ff.create_table(yearly_meet_counts)
-# fig.update_layout(title_text="Number of Meets per Year", title_x=0.5)
-# st.plotly_chart(fig)
-
-fig1 = px.bar(
-    yearly_meet_counts.reset_index(), x='Year', y='MeetName',
-    title='How many meets per year were there?',
-    labels={'MeetName': 'Number of Meets'},
-    color='MeetName',
-    color_continuous_scale=px.colors.sequential.Agsunset
-)
-fig1.update_layout(
-    title_font=dict(size=24),
-    xaxis_title='Year',
-    yaxis_title='Number of Meets'
-)
-st.plotly_chart(fig1)
-
-# how many lifters have competed in each year?
-yearly_lifter_counts = filtered_df_copy.groupby('Year')['Name'].count()
-print(yearly_lifter_counts)
-
-fig2 = px.bar(
-    yearly_lifter_counts.reset_index(), x='Year', y='Name',
-    title='How many lifters have competed in each year?',
-    labels={'Name': 'Total Participants'},
-    color='Name',
-    color_continuous_scale=px.colors.sequential.Agsunset
-)
-fig2.update_layout(
-    title_font=dict(size=24),
-    xaxis_title='Year',
-    yaxis_title='Total Participants'
-)
-st.plotly_chart(fig2)
-
-
-lifters_per_year = filtered_df_copy.groupby('Year')['Name'].nunique()
-print(lifters_per_year)
-
-fig3 = px.bar(
-    lifters_per_year.reset_index(),
-    x='Year',
-    y='Name',
-    title='How many Unique Lifters are there per Year?',
-    labels={'Name': 'Unique Lifters'},
-    color='Name',
-    color_continuous_scale=px.colors.sequential.Agsunset
-)
-fig3.update_layout(
-    title_font=dict(size=24),
-    xaxis_title='Year',
-    yaxis_title='Unique Lifters'
-)
-st.plotly_chart(fig3)
-
-# how many times have lifters competed each year?
-competitions_per_lifter_yearly = filtered_df_copy.groupby(['Year', 'Name']).size().reset_index(name='Competitions')
-print(competitions_per_lifter_yearly)
-
-times_competing_pa = competitions_per_lifter_yearly.groupby(['Year', 'Competitions']).size().unstack(fill_value=0)
-print(times_competing_pa)
-
-eight_comps_23 = competitions_per_lifter_yearly[(competitions_per_lifter_yearly['Year'] == 2023) & (competitions_per_lifter_yearly['Competitions'] == 8)]
-print(eight_comps_23)
-
-# has the gender distribution changed over time?
-gender_distribution_yearly = filtered_df_copy.groupby(['Year', 'Sex']).size().unstack(fill_value=0)
-gender_distribution_yearly['Total'] = gender_distribution_yearly.sum(axis=1)
-gender_distribution_yearly['% Male'] = ((gender_distribution_yearly['M'] / gender_distribution_yearly['Total']) * 100).round(0)
-gender_distribution_yearly['% Female'] = ((gender_distribution_yearly['F'] / gender_distribution_yearly['Total']) * 100).round(0)
-
-# display the updated DataFrame
-print(gender_distribution_yearly)
-
-# plot 4
-fig4 = px.bar(
-    gender_distribution_yearly.reset_index(),
-    x='Year',
-    y=['F', 'M'],
-    title='How has the gender balance of competitors changed over time?',
-    labels={'value': 'Number of Competitors', 'variable': 'Gender'},
-    barmode='group'
-)
-fig4.update_layout(
-    title_font=dict(size=24),
-    xaxis_title='Year',
-    yaxis_title='Number of Competitors',
-    legend_title='Gender'
-)
-st.plotly_chart(fig4)
 
 ######
 # streamlit widgets
